+++ pyvo/dal/tests/test_sia2.py
@@ -9,10 +9,11 @@ import requests_mock
 
 import pytest
 
-from pyvo.dal.sia2 import search, SIAService, SIAQuery
+from pyvo.dal.sia2 import search, SIA2Service, SIA2Query, SIAService, SIAQuery
 
 import astropy.units as u
 from astropy.utils.data import get_pkg_data_contents
+from astropy.utils.exceptions import AstropyDeprecationWarning
 
 get_pkg_data_contents = partial(
     get_pkg_data_contents, package=__package__, encoding='binary')
@@ -56,7 +57,7 @@ def test_search():
     _test_result(result)
 
 
-class TestSIAService:
+class TestSIA2Service:
 
     def test_capabilities(self):
         # this tests the SIA2 capabilities with various combinations:
@@ -79,20 +80,20 @@ class TestSIAService:
 
             # multiple interfaces with single security method each and
             # anonymous access.
-            service = SIAService('https://example.com/sia')
+            service = SIA2Service('https://example.com/sia')
             assert service.query_ep == 'https://example.com/sia/v2query'
 
             # one interfaces with BasicAA security method - this should fail
             # because pyvo does not support BasicAA
             with pytest.raises(AttributeError):
-                service = SIAService('https://example.com/sia-basicauth')
+                service = SIA2Service('https://example.com/sia-basicauth')
 
             # one interface with multiple security methods
-            service = SIAService('https://example.com/sia-newformat')
+            service = SIA2Service('https://example.com/sia-newformat')
             assert service.query_ep == 'https://example.com/sia/v2query'
 
             # multiple interfaces with single security method each (no anon)
-            service = SIAService('https://example.com/sia-priv')
+            service = SIA2Service('https://example.com/sia-priv')
             assert service.query_ep == 'https://example.com/sia/v2query'
 
     @pytest.mark.usefixtures('sia')
@@ -101,7 +102,7 @@ class TestSIAService:
     @pytest.mark.filterwarnings("ignore::astropy.io.votable.exceptions.W42")
     @pytest.mark.filterwarnings("ignore::astropy.io.votable.exceptions.W49")
     def test_search(self):
-        service = SIAService('https://example.com/sia')
+        service = SIA2Service('https://example.com/sia')
 
         positions = [
             (2, 4, 0.0166 * u.deg),
@@ -122,11 +123,18 @@ class TestSIAService:
         result = results[0]
         _test_result(result)
 
+        # test the deprecation
+        with pytest.warns(AstropyDeprecationWarning):
+            deprecated_service = SIAService('https://example.com/sia')
+            deprecated_results = deprecated_service.search(pos=positions)
+            result = deprecated_results[0]
+            _test_result(result)
 
-class TestSIAQuery():
+
+class TestSIA2Query():
 
     def test_query(self):
-        query = SIAQuery('someurl')
+        query = SIA2Query('someurl')
         query.field_of_view.add((10, 20))
         assert query['FOV'] == ['10.0 20.0']
         query.field_of_view.add((1 * u.rad, 60))
@@ -174,11 +182,24 @@ class TestSIAQuery():
         query.maxrec = 1000
         assert query['MAXREC'] == '1000'
 
-        query = SIAQuery('someurl', custom_param=23)
+        query = SIA2Query('someurl', custom_param=23)
         assert query['custom_param'] == ['23']
 
         query['custom_param'].append('-Inf 0')
         assert query['custom_param'] == ['23', '-Inf 0']
 
-        query = SIAQuery('someurl', custom_param=[('-Inf', 0), (2, '+Inf')])
+        query = SIA2Query('someurl', custom_param=[('-Inf', 0), (2, '+Inf')])
         assert query['custom_param'] == ['-Inf 0', '2 +Inf']
+
+        with pytest.warns(AstropyDeprecationWarning):
+            deprecated_query = SIAQuery('someurl')
+            deprecated_query.field_of_view.add((10, 20))
+            assert deprecated_query['FOV'] == ['10.0 20.0']
+
+
+def test_variable_deprecation():
+    # Test this while we are in the deprecation period, as the variable is durectly
+    # used at least by astroquery.alma
+    with pytest.warns(AstropyDeprecationWarning):
+        from pyvo.dal.sia2 import SIA_PARAMETERS_DESC
+        assert SIA_PARAMETERS_DESC