+++ CHANGES.rst
@@ -12,6 +12,12 @@
 - Removing defaults for optional SIAv1 and SSA query parameters to avoid
   unnecessarily overriding the server-side defaults. [#366]
 
+- Classes ``SIAService``, ``SIAQuery``, ``SIAResults`` for SIA v2 have been
+  renamed to ``SIA2Service``, ``SIA2Query``, ``SIA2Results`` respectively
+  as well as the variable ``SIA_PARAMETERS_DESC`` to
+  ``SIA2_PARAMETERS_DESC``. The old names now issue an
+  ``AstropyDeprecationWarning``. [#419]
+
 
 1.4 (2022-09-26)
 ================
@@ -73,7 +79,7 @@
 
 - Properly handle single string keywords value for regsearch(). [#201]
 
-- Add support for SIAv2. [#206]
+- Add support for SIA2. [#206]
 
 - Add kwargs to sia2. [#222]